requirements.txt
streamlit==1.44.1
pandas==2.2.3
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
openpyxl==3.1.5
